removed functions from order_type.go into order.go
